Move UnionFoldingBuilder into plugins
