+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/foldingbuilder/unionfoldingbuilder.kt
@@ -3,8 +3,10 @@ package arrow.meta.ide.plugins.proofs.foldingbuilder
 import arrow.meta.ide.IdeMetaPlugin
 import arrow.meta.phases.ExtensionPhase
 import com.intellij.lang.ASTNode
+import com.intellij.lang.folding.FoldingBuilderEx
 import com.intellij.lang.folding.FoldingDescriptor
 import com.intellij.openapi.editor.Document
+import com.intellij.psi.PsiElement
 
 fun IdeMetaPlugin.addUnionFoldingBuilder(): ExtensionPhase =
   addFoldingBuilder(
@@ -21,3 +23,18 @@ fun IdeMetaPlugin.addUnionFoldingBuilder(): ExtensionPhase =
       true
     }
   )
+
+class UnionFoldingBuilder : FoldingBuilderEx() {
+  override fun getPlaceholderText(node: ASTNode): String? =
+    "String | Int"
+
+  override fun buildFoldRegions(root: PsiElement, document: Document, quick: Boolean): Array<FoldingDescriptor> {
+    val result = arrayListOf<FoldingDescriptor>()
+    println("document=$document")
+    println("node=$root")
+    return result.toTypedArray()
+  }
+
+  override fun isCollapsedByDefault(node: ASTNode): Boolean =
+    true
+}
