+++ idea-plugin/src/main/resources/META-INF/plugin.xml
@@ -27,7 +27,7 @@
     <extensions defaultExtensionNs="com.intellij">
         <applicationInitializedListener implementation="arrow.meta.ide.MetaRegistrar"/>
         <lang.foldingBuilder language="kotlin"
-                             implementationClass="arrow.meta.ide.dsl.editor.foldingbuilder.UnionFoldingBuilder"/>
+                             implementationClass="arrow.meta.ide.plugins.proofs.foldingbuilder.UnionFoldingBuilder"/>
     </extensions>
 
     <extensions defaultExtensionNs="org.jetbrains.kotlin">
